Refactored some of the base test code to a single location in the atomix-testing module.
