+++ coordination/src/test/java/io/atomix/coordination/DistributedLeaderElectionTest.java
@@ -15,12 +15,14 @@
  */
 package io.atomix.coordination;
 
+import java.util.concurrent.atomic.AtomicLong;
+
+import org.testng.annotations.Test;
+
+import io.atomix.atomix.testing.AbstractAtomixTest;
 import io.atomix.coordination.state.LeaderElectionState;
 import io.atomix.copycat.client.CopycatClient;
 import io.atomix.resource.ResourceStateMachine;
-import org.testng.annotations.Test;
-
-import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Async leader election test.
@@ -28,7 +30,7 @@ import java.util.concurrent.atomic.AtomicLong;
  * @author <a href="http://github.com/kuujo">Jordan Halterman</a>
  */
 @Test
-public class DistributedLeaderElectionTest extends AbstractCoordinationTest {
+public class DistributedLeaderElectionTest extends AbstractAtomixTest {
 
   @Override
   protected ResourceStateMachine createStateMachine() {
