+++ coordination/src/test/java/io/atomix/coordination/DistributedLockTest.java
@@ -15,10 +15,12 @@
  */
 package io.atomix.coordination;
 
+import org.testng.annotations.Test;
+
+import io.atomix.atomix.testing.AbstractAtomixTest;
 import io.atomix.coordination.state.LockState;
 import io.atomix.copycat.client.CopycatClient;
 import io.atomix.resource.ResourceStateMachine;
-import org.testng.annotations.Test;
 
 /**
  * Async lock test.
@@ -26,7 +28,7 @@ import org.testng.annotations.Test;
  * @author <a href="http://github.com/kuujo">Jordan Halterman</a>
  */
 @Test
-public class DistributedLockTest extends AbstractCoordinationTest {
+public class DistributedLockTest extends AbstractAtomixTest {
 
   @Override
   protected ResourceStateMachine createStateMachine() {
